+++ packages/react-native/src/components/Card/Card.tsx
@@ -4,12 +4,11 @@ import {
   validateExpiry,
 } from '@evervault/card-validator';
 import { ReactNode, useEffect, useState } from 'react';
-import { useForm } from 'shared';
+import { useForm } from '../useForm';
 import { changePayload, isAcceptedBrand, isComplete } from './utilities';
-import type { CardForm, CardConfig } from './types';
+import type { CardForm, CardConfig, CardField, CardPayload } from './types';
 import { CardNumber } from './CardNumber';
 import { CardContext } from './context';
-import { CardField, CardPayload } from 'types';
 import { encrypt } from '../../sdk';
 import { CardCVC } from './CardCVC';
 import { CardHolder } from './CardHolder';