+++ packages/react-native/src/components/Card/index.ts
@@ -1,2 +1,1 @@
 export { Card, type CardProps } from './Card';
-export { CardPayload } from 'types';