+++ packages/react-native/src/components/Card/types.ts
@@ -1,4 +1,17 @@
-import type { CardBrandName } from 'types';
+export type CardBrandName =
+  | 'american-express'
+  | 'visa'
+  | 'mastercard'
+  | 'discover'
+  | 'jcb'
+  | 'diners-club'
+  | 'unionpay'
+  | 'maestro'
+  | 'mir'
+  | 'elo'
+  | 'hipercard'
+  | 'hiper'
+  | 'szep';
 
 export interface CardConfig {
   autoFocus?: boolean;
@@ -12,4 +25,29 @@ export interface CardForm {
   expiry: string;
 }
 
-export { CardBrandName };
+export type CardField = 'name' | 'number' | 'expiry' | 'cvc';
+
+export interface CardExpiry {
+  month: string | null;
+  year: string | null;
+}
+
+export interface CardPayload {
+  card: {
+    name: string | null;
+    brand: string | null;
+    localBrands: string[] | null;
+    number: string | null;
+    lastFour: string | null;
+    bin: string | null;
+    expiry: CardExpiry;
+    cvc: string | null;
+  };
+  isValid: boolean;
+  isComplete: boolean;
+  errors: null | Partial<{
+    number?: string;
+    cvc?: string;
+    expiry?: string;
+  }>;
+}