+++ e2e-tests/ui-components/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Patch Changes
 
+- @evervault/ui-components@1.3.1
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [e694b3b]
 - Updated dependencies [e694b3b]
   - @evervault/ui-components@1.3.0